UI changes

+++ frontend/src/components/common/ButtonReset.tsx
@@ -2,6 +2,7 @@ import ButtonWithModal, { ButtonWithModalProps } from './ButtonWithModal';
 import { fetchWithAuthAndParse } from '../../utils/auth.utils';
 import { useUser } from '../../hooks/useUser';
 import { UserScore } from '../../../../shared/types/dataTypes';
+import { useState } from 'react';
 
 export interface ButtonResetProps
   extends Omit<ButtonWithModalProps, 'onClick'> {
@@ -10,26 +11,36 @@ export interface ButtonResetProps
 
 export default function ButtonReset({ apiPath, ...props }: ButtonResetProps) {
   const { setUserScore } = useUser();
+  const [isLoading, setIsLoading] = useState(false);
 
   const handleReset = async () => {
-    const response = await fetchWithAuthAndParse<{ score: UserScore[] | null }>(
-      apiPath,
-      {
-        method: 'DELETE',
-      }
-    );
+    if (props.disabled || isLoading) {
+      return;
+    }
 
-    if (response) {
-      setUserScore(response?.score || null);
-    } else {
-      throw new Error('Reset failed');
+    setIsLoading(true);
+
+    try {
+      const response = await fetchWithAuthAndParse<{
+        score: UserScore[] | null;
+      }>(apiPath, {
+        method: 'DELETE',
+      });
+
+      if (response) {
+        setUserScore(response?.score || null);
+      } else {
+        throw new Error(`Reset failed: ${apiPath}`);
+      }
+    } finally {
+      setIsLoading(false);
     }
   };
 
   return (
     <ButtonWithModal
       modalMessage={props.modalMessage}
-      disabled={props.disabled}
+      disabled={props.disabled || isLoading}
       className={props.className}
       onClick={handleReset}
       successMessage="reset se zdařil"
